fish setup

+++ fish/config.fish
@@ -35,24 +35,28 @@ alias dl="docker ps -l -q"
 alias dip="docker inspect --format '{{ .NetworkSettings.IPAddress }}'"
 alias dexec="docker exec -i -t"
 
-if not test -f ~/.local/bin/fd
-	ln -s (which fdfind) ~/.local/bin/fd
-end
-
-## Exportiong other binaries to PATH
-set PATH $HOME/workspace/lua-language-server/bin $PATH
-
 set -gx FZF_DEFAULT_OPTS "--height 40% --layout=reverse --border --inline-info"
 set -gx FZF_DEFAULT_COMMAND "fdfind --hidden -E .git -E node_modules"
 set -gx FZF_CTRL_T_COMMAND "$FZF_DEFAULT_COMMAND"
- 
+
 mkdir -p ~/.config/fish/functions/
 echo fzf_key_bindings > ~/.config/fish/functions/fish_user_key_bindings.fish
 
 ## start asdf
-source ~/.asdf/asdf.fish
 if not test -f ~/.config/fish/completions/asdf.fish
 	mkdir -p ~/.config/fish/completions; 
 	ln -s ~/.asdf/completions/asdf.fish ~/.config/fish/completions
 end
 
+if test -f ~/.asdf/asdf.fish
+	source ~/.asdf/asdf.fish
+end
+
+if not test -f ~/.local/bin/fd 
+	and type -q fdfind
+	ln -s (which fdfind) ~/.local/bin/fd
+end
+
+if test -f ~/workspace/lua-language-server/bin
+	set -gx PATH ~/workspace/lua-language-server/bin $PATH
+end
